+++ meier/views/blog/routes/assets.py
